Code Refactoring after the review comments

+++ src/main/java/uk/co/policyexpert/supermarket/offers/Discount.java
@@ -1,7 +1,9 @@
 package uk.co.policyexpert.supermarket.offers;
 
+import java.math.BigDecimal;
+
 import uk.co.policyexpert.supermarket.CartItem;
 
 public interface Discount {
-	float applyDiscount(CartItem lineItem);
+	BigDecimal applyDiscount(CartItem lineItem);
 }
